requirements.txt
pdfplumber
openpyxl
pandas
python-docx
google-generativeai
PyPDF2